Add LocalAddr() to Incoming and Outgoing transfer interfaces.
This facilitates more detailed logging along with facilitating a
usecase I have which involves dynamically generating TFTP content
based on both the address of the system making the request and the
address that the host system is using to handle the request.  This is
useful in multi-homed environnments where we cannot assume that all of
the systems we act as a TFTP server for can see us at a single IP
address.