Refactor based on feedback from PR https://github.com/pin/tftp/pull/32
LocalAddr() renamed to LocalIP(), and will just return nil if we
cannot determine the actual IP used to recieve a packet from on the
server side.

Attempt to handle the case where we cannot open an IP control capable
PacketConn by calling back to using the UDPConn directly.

Change the signature of Serve() to return an error if we encounter an
uncorrectible setup error before entering the main packet handling
loop.