Make blksize negotiation a bit smarter.
If we are negotiating blksize, and acting as a server, do not just
agree with the blocksize the client wants us to use.  Instead, look up
the mtu of the interface we recieved the packet on, shave off
overhead, and if that is smaller than the requested blksize try to
negotiate that instead.  This is intended to work around embedded IP
stacks that do not calcualte their buffer size properly based on the
MTU that DHCP hands them, and is a lame attempt to work around
middleboxes that eat fragmented IP packets.

While we are at it, give Server a MaxBlockSize() method to match the
one the Client has.