Make the server side send packets from the same IP address it recieved them on.
If you have an interface with multiple IP addresses in the same subnet
bound to it, the server code opens a listen socket on all addresses to
communicate back to the client, which can cause the kernel to send
packets back to the client from a different IP address than the client
sent its initial request to.  Most of the commandline TFTP clients
don't care, but several of the ones embedded in NIC firmware do.  To
fix that, have the server side code open its client-specific listen
socket on the address it recieved the incoming request on.