Cope with clients using the broadcast address
Some clients use the broadcast address (255.255.255.255) to reach the
TFTP server in their network. This currently doesn't work because the
server then use this address as the source address in its responses.
But the broadcast address is not a valid source address and the
sendmsg call fail with an `invalid argument` error.

To cope with this check if the source address is the broadcast address
and if so replace it with 0.0.0.0 to use the OS default.

Signed-off-by: Alban Bedel <alban.bedel@aerq.com>